# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SAGE Platform API
  description: |
    RESTful API for the SAGE Platform - a project management and client collaboration system.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - Authentication endpoints: 5 requests per minute
    - General API: 100 requests per minute
    - File uploads: 10 requests per minute

    ## Error Responses
    All errors follow a standard format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable message",
        "details": {}
      }
    }
    ```
  version: 1.0.0
  contact:
    name: SAGE Support
    email: support@sage-platform.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: https://api.sage-platform.com/api
    description: Production
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List user's projects
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProjectStatus'
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjects'
  /projects/{id}:
    get:
      tags: [Projects]
      summary: Get project details
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Project with milestones and deliverables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Projects]
      summary: Update project status (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/ProjectStatus'
      responses:
        '200':
          description: Project updated

  # Milestones
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    ProjectStatus:
      type: string
      enum: [pending, active, in_progress, review, completed, on_hold, cancelled]
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        progress:
          type: number
        tierId:
          type: integer
        createdAt:
          type: string
          format: date-time
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/ProjectSummary'
        - type: object
          properties:
            milestones:
              type: array
              items:
                $ref: '#/components/schemas/Milestone'
            deliverables:
              type: array
              items:
                $ref: '#/components/schemas/Deliverable'
            payments:
              type: array
              items:
                $ref: '#/components/schemas/Payment'
    MilestoneStatus:
      type: string
      enum: [pending, in_progress, completed, blocked]
    Milestone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/MilestoneStatus'
        order:
          type: integer
        dueDate:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    Deliverable:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
          enum: [concept, draft, final, reference, other]
        fileUrl:
          type: string
        fileSize:
          type: integer
        mimeType:
          type: string
        createdAt:
          type: string
          format: date-time
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        stripePaymentId:
          type: string
        paidAt:
          type: string
          format: date-time
...
    PaginatedProjects:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProjectSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer